backend/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bleach==4.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2.11
django-cors-headers==4.3.1
django-dbbackup==4.1.0
django-filter==23.5
django-js-asset==2.2.0
django-modelcluster==6.3
django-mptt==0.14.0
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==20.0.4
html5lib==1.1
idna==3.7
importlib_metadata==7.1.0
l18n==2021.3
lxml==5.2.1
Markdown==3.6
mysqlclient==2.2.4
openpyxl==3.1.2
packaging==24.0
pillow==10.3.0
pillow_heif==0.16.0
psycopg==3.1.18
pypinyin==0.51.0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
typing_extensions==4.11.0
urllib3==2.2.1
wagtail==5.1.3
wagtail-headless-preview==0.8.0
wagtail-markdown==0.11.1
webencodings==0.5.1
Willow==1.6.3
zipp==3.18.1
